Enum numcodecs_zstd::ZstdCodecError

pub enum ZstdCodecError {
    HeaderEncodeFailed {
        source: ZstdHeaderError,
    },
    ZstdEncodeFailed {
        source: ZstdCodingError,
    },
    EncodedDataNotBytes {
        dtype: AnyArrayDType,
    },
    EncodedDataNotOneDimensional {
        shape: Vec<usize>,
    },
    HeaderDecodeFailed {
        source: ZstdHeaderError,
    },
    DecodeExcessiveEncodedData,
    DecodeProducedLess,
    ZstdDecodeFailed {
        source: ZstdCodingError,
    },
    MismatchedDecodeIntoArray {
        source: AnyArrayAssignError,
    },
}

Errors that may occur when applying the ZstdCodec.

Variants

HeaderEncodeFailed

ZstdCodec failed to encode the header

Fields

source: ZstdHeaderError

Opaque source error

ZstdEncodeFailed

ZstdCodec failed to encode the encoded data

Fields

source: ZstdCodingError

Opaque source error

EncodedDataNotBytes

ZstdCodec can only decode one-dimensional byte arrays but received an array of a different dtype

Fields

dtype: AnyArrayDType

The unexpected dtype of the encoded array

EncodedDataNotOneDimensional

ZstdCodec can only decode one-dimensional byte arrays but received an array of a different shape

Fields

shape: Vec<usize>

The unexpected shape of the encoded array

HeaderDecodeFailed

ZstdCodec failed to encode the header

Fields

source: ZstdHeaderError

Opaque source error

DecodeExcessiveEncodedData

ZstdCodec decode consumed less encoded data, which contains trailing junk

DecodeProducedLess

ZstdCodec produced less decoded data than expected

ZstdDecodeFailed

ZstdCodec failed to decode the encoded data

Fields

source: ZstdCodingError

Opaque source error

MismatchedDecodeIntoArray

ZstdCodec cannot decode into the provided array

Fields

source: AnyArrayAssignError

The source of the error