Function numcodecs_zstd::decompress

pub fn decompress(encoded: &[u8]) -> Result<AnyArray, ZstdCodecError>

Decompress the encoded data into an array using Zstandard.

Errors

Errors with

ZstdCodecError::HeaderDecodeFailed if decoding the header failed

ZstdCodecError::DecodeExcessiveEncodedData if the encoded data contains excessive trailing data junk

ZstdCodecError::DecodeProducedLess if decoding produced less data than expected

ZstdCodecError::ZstdDecodeFailed if an opaque decoding error occurred