numcodecs_zstd

Function decompress_into

pub fn decompress_into(
    encoded: &[u8],
    decoded: AnyArrayViewMut<'_>,
) -> Result<(), ZstdCodecError>

Decompress the encoded data into a decoded array using Zstandard.

Errors

Errors with

ZstdCodecError::HeaderDecodeFailed if decoding the header failed

ZstdCodecError::MismatchedDecodeIntoArray if the decoded array is of the wrong dtype or shape

ZstdCodecError::HeaderDecodeFailed if decoding the header failed

ZstdCodecError::DecodeExcessiveEncodedData if the encoded data contains excessive trailing data junk

ZstdCodecError::DecodeProducedLess if decoding produced less data than expected

ZstdCodecError::ZstdDecodeFailed if an opaque decoding error occurred